backend/requirements.txt
Flask==2.2.5
flask-cors==3.0.10
PyMySQL==1.0.3
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.0.1
